% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{fragment.estimation}
\alias{fragment.estimation}
\title{Predict fragment usage}
\usage{
fragment.estimation(output.dir, min.counts = 0, seq.depth = NULL,
  chr = "chr1", homozygote = FALSE, plot.width = 10, plot.height = 7)
}
\arguments{
\item{output.dir}{a directory for all the output. It is better to be a empty or non-existing directory;}

\item{min.counts}{the minimum couts cutoff}

\item{seq.depth}{the sequencing depth}

\item{chr}{the chromosome to estimate the sequencing depth}

\item{homozygote}{homozygote or heterozygote?}

\item{plot.width}{plot width, equal to "width" in pdf()}

\item{plot.height}{plot height, equal to "height" in pdf()}
}
\value{
a plot in pdf file: plot_fragment.pdf
}
\description{
Predict fragment usage
}
\examples{
test.data=system.file("extdata", "test.zip", package = "transgeneR")
temp <- tempdir()
unzip(test.data, exdir=temp)
output.dir=paste(temp,"/test",sep="")
fragment.estimation(output.dir, chr="chr1")
}
\references{
To be updated later
}
\author{
Guofeng Meng
}
