% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transgeneR.R
\name{transgeneR}
\alias{transgeneR}
\title{A integrated tool for transgenic integration discovery using sequencing data}
\usage{
transgeneR(output.dir, genome.ref, homozygote = FALSE, mate1 = NULL,
  mate2 = NULL, insert.seq = NULL, estimate.copy = FALSE,
  calculate.nonsplit.reads = TRUE, seq.depth = NULL,
  backgroud.chrom = "chr10", remove.duplicate = FALSE, min.clip = 30,
  max.tail = 10, min.split.reads = 2, cores = 1, max.fragment = 600,
  allow.gap = 100, homo.windows = 100, homo.step = 50, strict = 1,
  region = 10, span = 250, col.left = "lightblue",
  col.right = "lightgreen", max.show = 1000, gz = F, block.size = 2e+06,
  reads.dup.score = 50)
}
\arguments{
\item{output.dir}{a directory for all the output. It is better to be a empty or non-existing directory;}

\item{genome.ref}{the bowtie2 genome reference prefix. It should be "bt2_index_base" of the bowtie2-build;}

\item{homozygote}{the transgene sequence is integrated as homozygote? Default: False;}

\item{mate1}{the first fastq file of pair-end reads;}

\item{mate2}{the second fastq file of pair-end reads;}

\item{insert.seq}{the transgenic sequence in fasta format;}

\item{estimate.copy}{to estimate copy number or not. This option can be set TRUE only when whole genome sequencing data is available}

\item{calculate.nonsplit.reads}{to estimate number of nonsplit reads crossing the integration or rearrangement sites}

\item{seq.depth}{the sequencing depth of the whole genome sequencing data. It is used to estimated copy number of transgenic sequences. This option is used only when estimate.copy is TRUE}

\item{backgroud.chrom}{the chromosome used to estimate the sequencing information, e.g. coverage and copy number}

\item{remove.duplicate}{to remove the duplicated reads or not. Recommend to set it TURE if PCR-based sequencing data are used}

\item{min.clip}{the minimum clip size for reads split}

\item{max.tail}{the maximum tail unmapped length}

\item{min.split.reads}{the minimum number of reads for a validated integration site}

\item{cores}{the thread number}

\item{max.fragment}{equal to the setting of "-X" of bowtie2}

\item{allow.gap}{the maximum gap between two part of insertion}

\item{homo.windows}{the windows size during homologous sequence searching}

\item{homo.step}{the steps during homologous sequence searching}

\item{strict}{bool value, to use discordanantly mapped reads or not}

\item{region}{the tolarated region size of integration sites}

\item{span}{the spanned size to plot the integration sites}

\item{col.left}{the color of left part of integration site plot}

\item{col.right}{the color of right part of integration site plot}

\item{max.show}{the maximum number of read to display in integration site plot}

\item{gz}{bool value, to compressed the alignment output or not. Not recommend to set it TRUE}

\item{block.size}{the number of reads to analysis each run; it is used to control ROM usage}

\item{reads.dup.score}{the minimum alignment score difference for unique mapping}
}
\value{
The output are files located in "output.dir". They have a structure of:
output.dir:/
\itemize{
        \item aln_genome.sam (or aln_genome.sam.gz)
        \item aln_insert.sam (or aln_insert.sam.gz)
        \item assign.txt : read assignment in genome and transgenic sequence
        \item homo.txt : homologous annotation of transgeneic sequence
        \item mapping_summary.txt : reads alignment information
        \item report.txt : the predicted results
        \item warning.txt : the warning information
        \item plot_fragment.pdf the copy information of transgenic sequence
        \item insert_ref/ : the bowtie2 reference for transgenic sequence
        \item temp_files/: some temparory files
        \item sites/: the transgenic integration or recombination information
       \enumerate{
            \item site1.pdf: the plot for first site
            \item site2.pdf: the plot for second site
     }
}
}
\description{
A integrated tool for transgenic integration discovery using sequencing data
}
\details{
This a one-stop analysis pipeline to find the transgenic intergation sites and transgenic recombination. When the whole genome sequencing data are available, it can estimate the copy number of transgenic sequences in animal genome.  To do the whole analysis, it has following steps:
\itemize{
\item Build the bowtie2 reference for transgenic sequence. The output is store in a directory "insert_ref/";
\item Map the homologous regions of transgenic sequences in genome. Output is a file "homo.txt";
\item reads local alignment to both genome and transgenic sequences: in this step, users need to pre-install bowtie2 and build the genome reference of studied animals. This step will generate two file: aln_genome.sam and aln_insert.sam.
\item Assign the reads to genome, transgenic sequence or both and collect the clipping parts of read for second-round alignments. The output will be store in a dictory "temp_files/";
\item Second-round alignments. The output are stored as "temp_files/fragment_genome.sam" and "temp_files/fragment_insert.sam";
\item Connect the break sites in either transgenic sequence,  genome or between transgenic sequence and genome;
\item Make the plot for the split reads  in the integration sites. Figures are created in "sites/*.pdf";
\item If whole genome sequencing data are used, it can estimated both the full and imcomplete integration; and calculate their copy numbers. One figure is drawn as "plot_fragment.pdf"
}

In same case that users wish to re-run part of the analysis, user can just delete the output in mentioned steps and this will make it to skip the steps with outputs.
}
\examples{
#\dontrun{
test.data=system.file("extdata", "test.zip", package = "transgeneR")
temp <- tempdir()
unzip(test.data, exdir=temp)
output.dir=paste(temp,"/test",sep="")
transgene.sq=paste(temp,"/test/temp_files/insert.fa",sep="")
transgeneR(output.dir,  insert.seq=transgene.sq, estimate.copy=FALSE)
#}
}
\references{
To be updated later
}
\author{
Guofeng Meng
}
